Move argparse code into separate argparse modules
