+++ kcidb/monitor/spool/__init__.py
@@ -12,9 +12,10 @@ import datetime
 import email
 import email.policy
 from google.cloud import firestore
-from kcidb.misc import ArgumentParser, iso_timestamp, log_and_print_excepthook
+from kcidb.misc import log_and_print_excepthook
 from kcidb.monitor.misc import is_valid_firestore_id
 from kcidb.monitor.output import Notification
+from kcidb.argparse import ArgumentParser, iso_timestamp
 
 # Because we like the "id" name
 # pylint: disable=invalid-name
