+++ kcidb/mq/__init__.py
@@ -6,8 +6,6 @@ import json
 import logging
 import threading
 import sys
-import argparse
-import email
 import email.message
 import email.policy
 from abc import ABC, abstractmethod
@@ -17,7 +15,9 @@ import kcidb.io as io
 import kcidb.orm
 from kcidb import misc
 from kcidb.misc import LIGHT_ASSERTS
-
+import kcidb.argparse
+from kcidb.mq import argparse
+import kcidb.orm.argparse
 
 # Module's logger
 LOGGER = logging.getLogger(__name__)
@@ -598,178 +598,14 @@ class EmailSubscriber(Subscriber):
         self.parser = email.parser.Parser(policy=email.policy.SMTPUTF8)
 
 
-def argparse_add_args(parser):
-    """
-    Add common message queue arguments to an argument parser.
-
-    Args:
-        parser:     The parser to add arguments to.
-    """
-    parser.add_argument(
-        '-p', '--project',
-        help='ID of the Google Cloud project with the message queue',
-        required=True
-    )
-    parser.add_argument(
-        '-t', '--topic',
-        help='Name of the message queue topic',
-        required=True
-    )
-
-
-class ArgumentParser(misc.ArgumentParser):
-    """
-    Command-line argument parser with common message queue arguments added.
-    """
-
-    def __init__(self, *args, **kwargs):
-        """
-        Initialize the parser, adding common message queue arguments.
-
-        Args:
-            args:       Positional arguments to initialize ArgumentParser
-                        with.
-            kwargs:     Keyword arguments to initialize ArgumentParser with.
-        """
-        super().__init__(*args, **kwargs)
-        argparse_add_args(self)
-
-
-def argparse_publisher_add_args(parser, data_name):
-    """
-    Add message queue publisher arguments to an argument parser.
-
-    Args:
-        parser:     The parser to add arguments to.
-        data_name:  Name of the message queue data.
-    """
-    argparse_add_args(parser)
-    subparsers = parser.add_subparsers(dest="command",
-                                       title="Available commands",
-                                       metavar="COMMAND",
-                                       parser_class=argparse.ArgumentParser)
-    subparsers.required = True
-    parser.subparsers = {}
-    description = f"Initialize {data_name} publisher setup"
-    parser.subparsers["init"] = subparsers.add_parser(
-        name="init", help=description, description=description
-    )
-    description = \
-        f"Publish {data_name} from standard input, print publishing IDs"
-    parser.subparsers["publish"] = subparsers.add_parser(
-        name="publish", help=description, description=description
-    )
-    description = f"Cleanup {data_name} publisher setup"
-    parser.subparsers["cleanup"] = subparsers.add_parser(
-        name="cleanup", help=description, description=description
-    )
-
-
-class PublisherArgumentParser(misc.ArgumentParser):
-    """
-    Command-line argument parser with common message queue arguments added.
-    """
-
-    def __init__(self, data_name, *args, **kwargs):
-        """
-        Initialize the parser, adding common message queue arguments.
-
-        Args:
-            data_name:  Name of the message queue data.
-            args:       Positional arguments to initialize ArgumentParser
-                        with.
-            kwargs:     Keyword arguments to initialize ArgumentParser with.
-        """
-        super().__init__(*args, **kwargs)
-        self.subparsers = {}
-        argparse_publisher_add_args(self, data_name)
-
-
-def argparse_subscriber_add_args(parser, data_name):
-    """
-    Add message queue subscriber arguments to an argument parser.
-
-    Args:
-        parser:     The parser to add arguments to.
-        data_name:  Name of the message queue data.
-    """
-    argparse_add_args(parser)
-    parser.add_argument(
-        '-s', '--subscription',
-        help='Name of the subscription',
-        required=True
-    )
-    subparsers = parser.add_subparsers(dest="command",
-                                       title="Available commands",
-                                       metavar="COMMAND",
-                                       parser_class=argparse.ArgumentParser)
-    subparsers.required = True
-    parser.subparsers = {}
-
-    description = f"Initialize {data_name} subscriber setup"
-    parser.subparsers["init"] = subparsers.add_parser(
-        name="init", help=description, description=description
-    )
-
-    description = \
-        f"Pull {data_name} with a subscriber, print to standard output"
-    parser.subparsers["pull"] = subparsers.add_parser(
-        name="pull", help=description, description=description
-    )
-    parser.subparsers["pull"].add_argument(
-        '--timeout',
-        metavar="SECONDS",
-        type=float,
-        help='Wait the specified number of SECONDS for a message, '
-             'or "inf" for infinity. Default is "inf".',
-        default=math.inf,
-        required=False
-    )
-    parser.subparsers["pull"].add_argument(
-        '-m',
-        '--messages',
-        metavar="NUMBER",
-        type=misc.non_negative_int_or_inf,
-        help='Pull maximum NUMBER of messages, or "inf" for infinity. '
-             'Default is 1.',
-        default=1,
-        required=False
-    )
-
-    description = f"Cleanup {data_name} subscriber setup"
-    parser.subparsers["cleanup"] = subparsers.add_parser(
-        name="cleanup", help=description, description=description
-    )
-
-
-class SubscriberArgumentParser(misc.ArgumentParser):
-    """
-    Command-line argument parser with message queue subscriber arguments
-    added.
-    """
-
-    def __init__(self, data_name, *args, **kwargs):
-        """
-        Initialize the parser, adding message queue subscriber arguments.
-
-        Args:
-            data_name:  Name of the message queue data.
-            args:       Positional arguments to initialize ArgumentParser
-                        with.
-            kwargs:     Keyword arguments to initialize ArgumentParser with.
-        """
-        super().__init__(*args, **kwargs)
-        self.subparsers = {}
-        argparse_subscriber_add_args(self, data_name)
-
-
 def io_publisher_main():
     """Execute the kcidb-mq-io-publisher command-line tool"""
     sys.excepthook = misc.log_and_print_excepthook
     description = \
         'kcidb-mq-io-publisher - ' \
         'Kernel CI I/O data publisher management tool'
-    parser = PublisherArgumentParser("I/O data", description=description)
+    parser = argparse.PublisherArgumentParser("I/O data",
+                                              description=description)
     args = parser.parse_args()
     publisher = IOPublisher(args.project, args.topic)
     if args.command == "init":
@@ -793,8 +629,9 @@ def io_subscriber_main():
     description = \
         'kcidb-mq-io-subscriber - ' \
         'Kernel CI I/O data subscriber management tool'
-    parser = SubscriberArgumentParser("I/O data", description=description)
-    misc.argparse_output_add_args(parser.subparsers["pull"])
+    parser = argparse.SubscriberArgumentParser("I/O data",
+                                               description=description)
+    kcidb.argparse.output_add_args(parser.subparsers["pull"])
     args = parser.parse_args()
     subscriber = IOSubscriber(args.project, args.topic, args.subscription)
     if args.command == "init":
@@ -815,10 +652,11 @@ def pattern_publisher_main():
     description = \
         'kcidb-mq-pattern-publisher - ' \
         'Kernel CI ORM pattern publisher management tool'
-    parser = PublisherArgumentParser("ORM patterns", description=description)
+    parser = argparse.PublisherArgumentParser("ORM patterns",
+                                              description=description)
     parser.subparsers["publish"].add_argument(
         '--pattern-help',
-        action=kcidb.orm.query.PatternHelpAction,
+        action=kcidb.orm.argparse.PatternHelpAction,
         help='Print pattern string documentation and exit.'
     )
     args = parser.parse_args()
@@ -846,7 +684,8 @@ def pattern_subscriber_main():
     description = \
         'kcidb-mq-pattern-subscriber - ' \
         'Kernel CI ORM pattern subscriber management tool'
-    parser = SubscriberArgumentParser("ORM patterns", description=description)
+    parser = argparse.SubscriberArgumentParser("ORM patterns",
+                                               description=description)
     args = parser.parse_args()
     subscriber = ORMPatternSubscriber(args.project, args.topic,
                                       args.subscription)
@@ -870,7 +709,8 @@ def email_publisher_main():
     description = \
         'kcidb-mq-email-publisher - ' \
         'Kernel CI email queue publisher management tool'
-    parser = PublisherArgumentParser("email", description=description)
+    parser = argparse.PublisherArgumentParser("email",
+                                              description=description)
     args = parser.parse_args()
     publisher = EmailPublisher(args.project, args.topic)
     if args.command == "init":
@@ -888,7 +728,8 @@ def email_subscriber_main():
     description = \
         'kcidb-mq-email-subscriber - ' \
         'Kernel CI email queue subscriber management tool'
-    parser = SubscriberArgumentParser("email", description=description)
+    parser = argparse.SubscriberArgumentParser("email",
+                                               description=description)
     args = parser.parse_args()
     subscriber = EmailSubscriber(args.project, args.topic, args.subscription)
     if args.command == "init":
