+++ kcidb/orm/__init__.py
@@ -5,7 +5,6 @@ objects, but without the object-oriented interface.
 
 import logging
 from abc import ABC, abstractmethod
-import kcidb.misc
 from kcidb.misc import LIGHT_ASSERTS
 from kcidb.orm import query, data
 
@@ -271,60 +270,3 @@ class Cache(Source):
         }
         assert LIGHT_ASSERTS or data.SCHEMA.is_valid(response)
         return response
-
-
-def argparse_add_args(parser):
-    """
-    Add common ORM arguments to an argument parser.
-
-    Args:
-        The parser to add arguments to.
-    """
-    parser.add_argument(
-        "pattern_strings",
-        nargs="*",
-        default=[],
-        metavar="PATTERN",
-        help="Object-matching pattern. "
-        "See pattern documentation with --pattern-help.",
-    )
-    parser.add_argument(
-        "--pattern-help",
-        action=query.PatternHelpAction,
-        help="Print pattern string documentation and exit.",
-    )
-
-
-class ArgumentParser(kcidb.misc.ArgumentParser):
-    """
-    Command-line argument parser with common ORM arguments added.
-    """
-
-    def __init__(self, *args, **kwargs):
-        """
-        Initialize the parser, adding common ORM arguments.
-
-        Args:
-            args:   Positional arguments to initialize ArgumentParser with.
-            kwargs: Keyword arguments to initialize ArgumentParser with.
-        """
-        super().__init__(*args, **kwargs)
-        argparse_add_args(self)
-
-
-class OutputArgumentParser(kcidb.misc.OutputArgumentParser):
-    """
-    Command-line argument parser for tools outputting JSON,
-    with common ORM arguments added.
-    """
-
-    def __init__(self, *args, **kwargs):
-        """
-        Initialize the parser, adding JSON output arguments.
-
-        Args:
-            args:   Positional arguments to initialize ArgumentParser with.
-            kwargs: Keyword arguments to initialize ArgumentParser with.
-        """
-        super().__init__(*args, **kwargs)
-        argparse_add_args(self)
